update Gemfile.lock
I created a new bot in Discord Developer Portal, run it and met an error about the privileged intents.
But the message content intent is allowed in Portal.
Finally, the error is resolved by this code modification.
Is the error caused by the discordrb update, or the new bot?
That is a mystery.